+++ src/constants.ts
@@ -27,3 +27,9 @@ export const staticPathsFetchBatch = 50;
 export const staticPathsFallback = (exportMode
   ? false
   : process.env.NEXT_PUBLIC_STATIC_PATHS_FALLBACK) as boolean | "blocking";
+
+export const paymentGatewayNames = {
+  dummy: "mirumee.payments.dummy",
+  adyen: "mirumee.payments.adyen",
+  stripe: "saleor.payments.stripe",
+};